web/src/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.8
django-admin-list-filter-dropdown==1.0.3
django-allauth==0.58.2
django-autoslug==1.9.9
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
django-cities-light==3.10.0
django-crispy-forms==2.1
django-filter==23.5
django-model-utils==4.3.1
django-multiselectfield==0.1.12
django-table-sort==0.5.0
django-tables2==2.7.0
idna==3.6
numpy==1.24.4
oauthlib==3.2.2
pandas==2.0.3
pillow==10.2.0
progressbar2==4.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-utils==3.8.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
swapper==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
Unidecode==1.3.7
urllib3==2.1.0
XlsxWriter==3.1.9
